Add files via upload
the hospitalization rate per age group (old variant) is stored as the RDS file ifr_vec_CAN.
The death rate per age group (old variant) is stored as the RDS file ihr_vec_CAN.
Task: can we update/ replace these two files to reflect the current hospitalization and death rate of the new variant?

yll_vec_CAN file is for long Covid symtomp. We don't model this
tupper_QALY_dollars  file is for the reduction in quality of life, measured in dollars. We also don't model this.

The mu_... files are for building the contact matrix. This is for essential worker groups. This was essentially disabled when I grouped the essential worker groups to their corresponding normal age group in Scenario 1. Hence we don't need to worry about these.